services/order_service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy[asyncio]==2.0.30
aiomysql==0.2.0
pydantic==2.7.4
pydantic-settings==2.3.4
redis[hiredis]==5.0.7
aiokafka==0.11.0
python-jose[cryptography]==3.3.0
